requirements_deployment.txt
# Core Streamlit and Web Framework
streamlit==1.28.1
streamlit-webrtc==0.47.1

# Database and Data Processing
pandas==2.1.3
numpy==1.24.3
sqlite3
psycopg2-binary==2.9.7
sqlalchemy==1.4.46

# Face Recognition and Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
Pillow==10.1.0

# Excel and File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Time and Date Handling
pytz==2023.3
python-dateutil==2.8.2

# Cloud Storage (for deployment)
boto3==1.34.0
google-cloud-storage==2.10.0

# API and Web Services
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0

# Google Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Data Visualization
plotly==5.17.0
matplotlib==3.8.2

# Security and Encryption
cryptography==41.0.7
bcrypt==4.1.2

# Logging and Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# System Utilities
python-dotenv==1.0.0
click==8.1.7